make schema field typescript, remove utilities class and single use
