+++ packages/compass-schema/package.json
@@ -73,6 +73,7 @@
     "@mongodb-js/webpack-config-compass": "^1.0.4",
     "@testing-library/react": "^12.1.4",
     "@types/chai": "^4.2.21",
+    "@types/lodash.find": "^4.6.7",
     "@types/mocha": "^9.0.0",
     "@types/react": "^17.0.5",
     "@types/react-dom": "^17.0.10",
