--- packages/compass-schema/src/components/field/index.js
@@ -1,2 +1,0 @@
-import Field from './field';
-export default Field;
